+++ src/datasources/alarms/components/editors/alarms-trend/AlarmsTrendQueryEditor.tsx
@@ -2,10 +2,11 @@ import { AlarmsTrendQuery } from 'datasources/alarms/types/AlarmsTrend.types';
 import React, { useEffect, useState } from 'react';
 import { InlineField } from 'core/components/InlineField';
 import { AlarmsQueryBuilder } from '../../query-builder/AlarmsQueryBuilder';
-import { ERROR_SEVERITY_WARNING, LABEL_WIDTH, labels, tooltips } from 'datasources/alarms/constants/AlarmsQueryEditor.constants';
+import { ERROR_SEVERITY_WARNING, LABEL_WIDTH, labels, SECONDARY_LABEL_WIDTH, tooltips } from 'datasources/alarms/constants/AlarmsQueryEditor.constants';
 import { AlarmsTrendQueryHandler } from 'datasources/alarms/query-type-handlers/alarms-trend/AlarmsTrendQueryHandler';
 import { Workspace } from 'core/types';
 import { FloatingError } from 'core/errors';
+import { InlineSwitch, Stack } from '@grafana/ui';
 
 type Props = {
   query: AlarmsTrendQuery;
@@ -36,20 +37,36 @@ export function AlarmsTrendQueryEditor({ query, handleQueryChange, datasource }:
     }
   };
 
+  const onGroupBySeverityChange = (isGroupBySeverityChecked: boolean) => {
+    handleQueryChange({ ...query, groupBySeverity: isGroupBySeverityChecked });
+  }
+
   return (
     <>
-      <InlineField
-        label={labels.queryBy}
-        labelWidth={LABEL_WIDTH}
-        tooltip={tooltips.queryBy}
-      >
-        <AlarmsQueryBuilder
-          filter={query.filter}
-          globalVariableOptions={datasource.globalVariableOptions()}
-          workspaces={workspaces}
-          onChange={onFilterChange}
-        />
-      </InlineField>
+      <Stack>
+        <InlineField
+          label={labels.queryBy}
+          labelWidth={LABEL_WIDTH}
+          tooltip={tooltips.queryBy}
+        >
+          <AlarmsQueryBuilder
+            filter={query.filter}
+            globalVariableOptions={datasource.globalVariableOptions()}
+            workspaces={workspaces}
+            onChange={onFilterChange}
+          />
+        </InlineField>
+        <InlineField
+          label={labels.groupBySeverity}
+          labelWidth={SECONDARY_LABEL_WIDTH}
+          tooltip={tooltips.groupBySeverity}
+        >
+          <InlineSwitch
+            onChange={event => onGroupBySeverityChange(event.currentTarget.checked)}
+            value={query.groupBySeverity}
+          />
+        </InlineField>
+      </Stack>
       <FloatingError
         message={datasource.errorTitle}
         innerMessage={datasource.errorDescription}