+++ src/datasources/alarms/types/types.ts
@@ -86,3 +86,11 @@ export enum AlarmTransitionType {
   Clear = 'CLEAR',
   Set = 'SET',
 };
+
+export enum AlarmTransitionSeverityLevel {
+  Clear = -1,
+  Low = 1,
+  Moderate = 2,
+  High = 3,
+  Critical = 4,
+}